requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Faker==25.8.0
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
humanize==4.9.0
Jinja2==3.1.4
passlib==1.7.4
python-dateutil==2.9.0.post0
SQLAlchemy==2.0.30
thefuzz==0.22.1
WTForms==3.1.2
